wip(failed): WOR-282 pre-failure state — retry may resume from here
