+++ tests/test_watcher_finalize_retry.py
@@ -1,0 +1,140 @@
+"""Tests for finalize_worker — retry wiring and safe_set_state resilience."""
+
+from __future__ import annotations
+
+import subprocess
+from pathlib import Path
+from unittest.mock import MagicMock, patch
+
+from app.core.linear_client import LinearError
+from app.core.watcher.watcher_types import ActiveWorker
+from tests._finalize_helpers import _call_finalize
+from tests.conftest import make_manifest
+
+# ---------------------------------------------------------------------------
+# retry_count wiring
+# ---------------------------------------------------------------------------
+
+
+def test_finalize_worker_retry_count_zero_on_success(tmp_path: Path) -> None:
+    manifest = make_manifest(ticket_id="WOR-10", worker_branch="wor-10-test-ticket")
+    metrics_mock = MagicMock()
+    worker = ActiveWorker(
+        ticket_id="WOR-10",
+        linear_id="fake-linear-id",
+        manifest=manifest,
+        worktree_path=tmp_path,
+        process=MagicMock(spec=subprocess.Popen),
+    )
+    with (
+        patch("app.core.watcher.watcher_finalize.run_checks", return_value=(True, [])),
+        patch(
+            "app.core.watcher.watcher_finalize.create_pr",
+            return_value="https://github.com/example/pr/1",
+        ),
+        patch("app.core.watcher.watcher_finalize.cleanup_worktree"),
+    ):
+        _call_finalize(worker, metrics=metrics_mock)
+
+    call_kwargs = metrics_mock.record.call_args[0][0]
+    assert call_kwargs.retry_count == 0
+
+
+def test_finalize_worker_retry_count_increments_on_check_failure(
+    tmp_path: Path,
+) -> None:
+    manifest = make_manifest(ticket_id="WOR-10", worker_branch="wor-10-test-ticket")
+    worker = ActiveWorker(
+        ticket_id="WOR-10",
+        linear_id="fake-linear-id",
+        manifest=manifest,
+        worktree_path=tmp_path,
+        process=MagicMock(spec=subprocess.Popen),
+    )
+    with (
+        patch("app.core.watcher.watcher_finalize.run_checks", return_value=(False, [])),
+        patch("app.core.watcher.watcher_finalize.cleanup_worktree"),
+    ):
+        _call_finalize(worker)
+        _call_finalize(worker)
+
+    assert worker.retry_count == 2
+
+
+def test_finalize_worker_retry_count_two_failures_then_success(
+    tmp_path: Path,
+) -> None:
+    manifest = make_manifest(ticket_id="WOR-10", worker_branch="wor-10-test-ticket")
+    metrics_mock = MagicMock()
+    worker = ActiveWorker(
+        ticket_id="WOR-10",
+        linear_id="fake-linear-id",
+        manifest=manifest,
+        worktree_path=tmp_path,
+        process=MagicMock(spec=subprocess.Popen),
+    )
+    check_results = [(False, []), (False, []), (True, [])]
+    with (
+        patch(
+            "app.core.watcher.watcher_finalize.run_checks", side_effect=check_results
+        ),
+        patch(
+            "app.core.watcher.watcher_finalize.create_pr",
+            return_value="https://github.com/example/pr/1",
+        ),
+        patch("app.core.watcher.watcher_finalize.cleanup_worktree"),
+    ):
+        _call_finalize(worker, metrics=metrics_mock)
+        _call_finalize(worker, metrics=metrics_mock)
+        _call_finalize(worker, metrics=metrics_mock)
+
+    call_kwargs = metrics_mock.record.call_args[0][0]
+    assert call_kwargs.retry_count == 2
+
+
+# ---------------------------------------------------------------------------
+# safe_set_state — daemon survives LinearError at finalize set_state sites
+# ---------------------------------------------------------------------------
+
+
+def test_finalize_worker_set_state_failure_nonzero_no_crash(tmp_path: Path) -> None:
+    manifest = make_manifest(ticket_id="WOR-10", worker_branch="wor-10-test-ticket")
+    linear_mock = MagicMock()
+    linear_mock.set_state.side_effect = LinearError("rate limit")
+
+    worker = ActiveWorker(
+        ticket_id="WOR-10",
+        linear_id="fake-linear-id",
+        manifest=manifest,
+        worktree_path=tmp_path,
+        process=MagicMock(spec=subprocess.Popen),
+    )
+
+    with patch("app.core.watcher.watcher_finalize.cleanup_worktree"):
+        _call_finalize(worker, returncode=1, linear=linear_mock)
+
+
+def test_finalize_worker_set_state_failure_success_path_no_crash(
+    tmp_path: Path,
+) -> None:
+    manifest = make_manifest(ticket_id="WOR-10", worker_branch="wor-10-test-ticket")
+    linear_mock = MagicMock()
+    linear_mock.set_state.side_effect = LinearError("network error")
+
+    worker = ActiveWorker(
+        ticket_id="WOR-10",
+        linear_id="fake-linear-id",
+        manifest=manifest,
+        worktree_path=tmp_path,
+        process=MagicMock(spec=subprocess.Popen),
+    )
+
+    with (
+        patch("app.core.watcher.watcher_finalize.run_checks", return_value=(True, [])),
+        patch(
+            "app.core.watcher.watcher_finalize.create_pr",
+            return_value="https://github.com/example/pr/1",
+        ),
+        patch("app.core.watcher.watcher_finalize.cleanup_worktree"),
+    ):
+        _call_finalize(worker, linear=linear_mock)
